fix(request): wait for token selector before preparing tx
Because the token selector and the request component both depend on
selectedTokenAddress and selectedChainID, they will try to execute
their effects at the same time. This causes the request component to use
incorrect token price data. This commit fixes that by adding a
tokenPriceCompleted state to the token selector context, which is set to
true once the token price has been fetched.

Refs: #98