refactor(request): treat token data from selector as an unit
From the request pay init, we need to now the token address, chain and
decimals in order to calculate the squid router fee. Previously we
depended on those three data points separately. This caused that when
the user seleted a token, we tried to calculate the route for that token
but with the decimals from the previous one (the effect tasked with
cleaning this also depends on the change of selected token, so they both
executed at the same time).

In reality, from the pay we depend on the whole token data and only care
when it updates completely. This refactor reflects that and avoid
calling squid with bad data.

Refs:  #102