fix: token amount input
Refactor token amount input to use a more robust approach.

The component communicates two values: the token amount and the USD value. Also keeps two display values internally that it shows: the main display amount and the alternative amount, which depending on the mode that the input is using will be between token and usd or fiat and usd